* Fix the function signature for filent.c's getmilliseconds.

+++ src/filent.c
@@ -713,7 +713,7 @@ void getprocesspath(char* buffer, size_t bufferLen)
 
 #pragma comment(lib, "winmm.lib")
 
-unsigned int getmilliseconds()
+unsigned long long getmilliseconds()
 {
     return timeGetTime();
 }
